requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
alpha_vantage==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
finnhub-python==2.4.23
Flask==3.1.0
frozendict==2.4.6
frozenlist==1.6.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.3
multitasking==0.0.11
narwhals==1.35.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.7
plotly==6.0.1
propcache==0.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.0
yfinance==0.2.55
